requirements.txt
playwright
requests
moviepy
google-auth-oauthlib
google-api-python-client
